LeetCode: Search in Rotated Sorted Array II
Approach 1: The key is to figuring out which section is increasing.
Think about these three cases: 1. A[l] < A[mid];
2. A[l] > A[mid]; 3. A[l] = A[mid].
For case 1, A[l] to A[mid] is incrementing. If target >= A[l] && target
< A[mid], check the section A[l]..A[mid-1]. Else target < A[l] ||
target > A[mid], then check the section A[mid+1]..A[r].
For case 2, A[mid] to A[r] is incrementing. If target <= A[r] && target
> A[mid], check the section A[mid+1]..A[r]. Else target >A[r] || target
< A[mid], check the section A[l]..A[mid-1].
For case 3, there are duplicates, A[l]==A[mid]. Move l to skip the
duplicates.
The worst case is O(n). Average is O(lgn).