LeetCode: Bitwise AND of Numbers Range
Approach 1: when m != n, there must be at least one odd number and one
even number, that means the last bit of the result is 0. Shift m and n
one bit to the right, until they are the same, and count how many times
they are shifted. the reset of m*how many times they are shifted is the
result.