LeetCode: ZigZag Conversion
Approach 1:
numRows=5
   0         1            2
0 A         I             Q
1 B     H J         P  R
2 C   G   K     O     S
3 D F      L  N        T
4 E         M            U

Consider the column with numbers as the main column.
For row i, column j, the index of the char in S is (2*numRows-2)*j+i.
And every element in the main column, there may be a char follow it,
for example, row 1, column 0, main element is B, followed by H.
The non-main element index is the index of main element +
(numRows-i-1)*2.