LeetCode: Longest Palindromic Substring
Approach 3: for every element in the input, make it as the center of a
possible palindrome, from the center to left and right, find the
possible palindrome. Then assume the current element and the element
next to current element as the center of a palindrome, and move to left
and right, to find the possible palindrome.